infra/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trading-simulator-infra"
version = "0.1.0"
description = "AWS CDK (Python) infrastructure for the trading simulator"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Trading Simulator", email = "example@example.com" },
]

dependencies = [
  "aws-cdk-lib>=2.160.0,<3.0.0",
  "constructs>=10.3.0,<11.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
# This project is an application (CDK app), not a library. Explicitly
# list the files that should go into the wheel so hatchling can build it.
include = [
  "app.py",
  "simulator_stack.py",
]



